security sweep: 9 critical contract/RPC fixes + 43/43 matrix pass\n\nContract fixes:\n- CON-01: moltoracle staleness 3600→9000 slots (1hr at 400ms/slot)\n- CON-02: shielded_pool reentrancy guards\n- CON-03: shielded_pool admin caller verification\n- CON-04: shielded_pool pause/unpause mechanism\n- CON-05: clawpump reject unconfigured MOLT (return false)\n- CON-06: lobsterlend u128 cast for health factor overflow\n- CON-07: moltdao PROPOSAL_SIZE 210→212\n- DEX-02: dex_router→AMM cross-call (action byte + full args)\n- GX-02: tx_to_rpc_json status rationale documented\n- GX-03: MoltCoin supply aligned 1M→1B\n- GX-04: mint() documentation (native vs wrapper layer)\n\nTests:\n- test-critical-security.sh: 40/40 security checks\n- test-mkt-featured-filter.sh: T-MKT-008 coverage\n- Matrix: 43/43 pass, cargo: 1085/1085 pass"

+++ contracts/lobsterlend/src/lib.rs
@@ -744,10 +744,12 @@ pub extern "C" fn get_account_info(user_ptr: *const u8) -> u32 {
     let borrow = compute_current_borrow(&hex);
 
     // Health factor in basis points (10000 = 1.0)
+    // AUDIT-FIX CON-06: Cast to u128 to prevent overflow for large deposits
+    // (deposit * 8500 overflows u64 when deposit > ~2.17×10¹⁵ shells ≈ 2.17M MOLT)
     let health_factor = if borrow == 0 {
         u64::MAX // Infinite health
     } else {
-        deposit * LIQUIDATION_THRESHOLD_PERCENT * 100 / borrow
+        ((deposit as u128) * (LIQUIDATION_THRESHOLD_PERCENT as u128) * 100 / (borrow as u128)) as u64
     };
 
     let mut result = Vec::with_capacity(24);
